Tag the RosBE-Windows 1.3.0 release

+++ reactos/lib/sdk/crt/stdlib/errno.c
@@ -13,9 +13,9 @@
 /*
  * @implemented
  */
-unsigned long* __doserrno(void)
+int* __doserrno(void)
 {
-  return (unsigned long*)(&GetThreadData()->tdoserrno);
+  return (int*)(&GetThreadData()->tdoserrno);
 }
 
 /*
